+++ cpucontrolsection.h
